Add function to logout an OAuth user
Adds a line to the OAuth login Sapling to set a local session storage
value to Splinter REST API's `/oauth/logout` endpoint. Also adds a
function to the Profile Sapling to check if this value has been set, and
if so, makes a request to the logout endpoint with the correct
authorization headers.

Signed-off-by: Shannyn Telander <telander@bitwise.io>